# api/api.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: ElektroGo
  description: "Documentació OpenAPI per a l'API REST desenvolupada a partir de l'aplicació **ElektroGo**. <br> Per tal de poder executar les operacions cal estar conectat a la VPN de la UPC. Totes les operacions requereixen d'una api_key. En cas de no tenir-ne la podeu solicitar a qualsevol membre de l'equip."
servers:
  - url: 'http://10.4.41.58:8080/'
paths:
  /users:
    get:
      tags:
      - Users
      summary: Obté tots els usuaris.
      operationId: getAllUsers
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/User'
    post:
      tags:
      - Users
      summary: Crea un usuari.
      operationId: postUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        201:
          description: successful operation
        438:
          description: 'Error: User Alredy Exists'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error438'
  /users/{username}:
    get:
      tags:
      - Users
      summary: Obté un usuari donat un username.
      operationId: getUserByUsername
      parameters:
      - name: username
        in: path
        description: Username de l'usuari
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        437:
          description: 'Error: User Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error437'
    delete:
      tags:
      - Users
      summary: Esborra un usuari.
      operationId: deleteUser
      parameters:
      - name: username
        in: path
        description: Username de l'usuari
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
        437:
          description: 'Error: User Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error437'
  /users/provider/{provider}/id/{id}:
    get:
      tags:
      - Users
      summary: Obté un usuari donat un parell (provider, id).
      operationId: getUseryId
      parameters:
      - name: provider
        in: path
        description: Proveïdor de login de l'usuari
        required: true
        schema:
          type: string
      - name: id
        in: path
        description: ID de l'usuari pel proveïdor de login
        required: true
        schema:
          type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        437:
          description: 'Error: User Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResultElektroGo'
              examples:
                errorExample:
                  $ref: '#/components/examples/error437'
...
  /car-pooling/{id}/users:
    get:
      tags:
      - Car Pooling
      summary: Obté tots els usuaris d'un trajecte de carpooling.
      operationId: getCarPoolingUsers
      parameters:
      - name: id
        in: path
        description: ID del trajecte de carpooling.
        required: true
        schema:
          type: integer
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items: 
                  $ref: '#/components/schemas/User'
        404:
          description: 'Error: Trip Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorExample:
                  $ref: '#/components/examples/error404'
...
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
          example: "samito"
        email:
          type: string
          example: "samito33@gmail.com"
        id:
          type: string
          example: "104069839557168305566"
        provider:
          type: string
          example: "GOOGLE"
        name:
          type: string
          example: "Samuel Daniel Foster"
        givenName:
          type: string
          example: "Samuel"
        familyName:
          type: string
          example: "Daniel Foster"
        imageUrl:
          type: string
          example: "https://lh3.googleusercontent.com/a-/AOh14GiIcAvvngHoDsa9CP5KmAgA1CfJglhCJjnjJR-z"
...
    ErrorResult:
      type: object
      properties:
        status:
          type: integer
          enum: [400, 401, 404]
        error: 
          type: string
          enum: ["Bad Request", "Unauthorized", "Not found"]
        message: 
          type: string
    ErrorResultElektroGo:
      type: object
      properties:
        timestamp:
          type: string
        status:
          type: integer
        error:
          type: string
        path:
          type: string
  examples:
...
    error404:
      value:
        status: 404
        error: "Not found"
        message: "There are no trips in the radius specified"
...
    error437:
      value:
        timestamp: "2022-05-13T21:41:52.369+00:00"
        status: 437
        error: "Http Status 437"
        path: "/users"
    error438:
      value:
        timestamp: "2022-05-13T21:41:52.369+00:00"
        status: 438
        error: "Http Status 438"
        path: "/users"
...